For testability without PyQt, move most code into utility library
Adds one example test
